% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.sigs.R
\name{compare.sigs}
\alias{compare.sigs}
\title{Compares sparse matrixes of gene signatures}
\usage{
compare.sigs(sig.matrix, test.matrix)
}
\arguments{
\item{sig.matrix}{Signature sparse matrix}

\item{test.matrix}{Test signatures sparse matrix}
}
\value{
A data table of the comparisons of sig.matrix and test.matrix
  including odds ratios, p.values and FDR adjusted q.values
}
\description{
Compares sparse matrixes of gene signatures
}
\examples{
compare.sigs(sig.matrix, test.matrix)
}
