% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sigs.to.matrix.R
\name{convert.sigs.to.matrix}
\alias{convert.sigs.to.matrix}
\title{Converts signatures into a sparse matrix}
\usage{
convert.sigs.to.matrix(signatures, genes)
}
\arguments{
\item{signatures}{Signatures table from loadSig}

\item{genes}{A list of all the genes that will be used in the comparison}
}
\value{
A sparse matrix of signatures by gene
}
\description{
Converts signatures into a sparse matrix
}
\examples{
convert.sigs.to.matrix(signatures, genes)
}
