% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES.plot.R
\name{ES.plot}
\alias{ES.plot}
\title{Enrichment set plot}
\usage{
ES.plot(gmt = NULL, cls = NULL, comparison = NULL, geneset = NULL,
  sn.table = NULL, msigdb = NULL)
}
\arguments{
\item{gmt}{A gmt formatted gene expression data set}

\item{cls}{A cls formatted classifier}

\item{comparison}{Which levels of cls should be compared defaults to levels 1 and 2}

\item{geneset}{Either a vector of gene identifiers or a geneset name in combination with msigdb parameter}

\item{sn.table}{(optional) A signal to noise named vector can be used in place of gmt, cls and comparison}

\item{msigdb}{(optional) A database of signatures}
}
\description{
Enrichment set plot
}
