% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FES.plot.R
\name{FES.plot}
\alias{FES.plot}
\title{FES plot}
\usage{
FES.plot(gmt = NULL, cls = NULL, comparison = NULL, geneset = NULL,
  sn.table = NULL, msigdb = NULL)
}
\arguments{
\item{gmt}{a gmt formatted gene expression data.frame}

\item{cls}{a vector describing the classes the gmt datasets belong to}

\item{comparison}{which comparison of classes is desired defaults to the
first two levels from cls}

\item{geneset}{the set of genes to test and plot}

\item{sn.table}{the ordered, valued ranking of genes}

\item{msigdb}{a database of signatures}
}
\value{
A ggplot2 plot of the FES score in a style similar to the GSEA style
  plots
}
\description{
FES(Fast Enrichment Score or Fisher Enrichment Score) uses a different method
of calculating the enrichment score as compared with GSEA.  In FES, the
hypergeometric distribution is used to determine the p-value for the
partitioning of members of the geneset within the ordered valued gene
ranking.  Two different p-values are calculated from the opposite ends of the
ranking to measure enrichment at either the high or low ends of the spectrum.
The use of the hypergeometric distribution rather than the complicated method
used in GSEA means this method is very fast and yields results which are
similar to GSEA itself.  In comparison to GSEA, FES is more sensitive to
concentration of geneset members in the middle of the ranking as happens with
GSEA v1 and is avoided by GSEA v2.  All results from FES or GSEA should be
evaluated by looking at the plots and changes in gene expression to evaluate
the feasibility of detecting the shift in the gene set distribution using
experimental tools.  Typically if the changes are smaller than a 0.5 on the
log2FC scale, they will be challenging to detect by methods such as Q-PCR.
}
