% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FES.R
\name{FES}
\alias{FES}
\title{FES}
\usage{
FES(sn.table, geneset, hits.only = TRUE)
}
\arguments{
\item{sn.table}{a table of ordered, valued gene rankings}

\item{geneset}{a set of genes to determine the FES score}

\item{hits.only}{Emit all genes(False) or only those which are in the geneset(True)}
}
\value{
A data.frame of FES scores for the selected genes.
}
\description{
FES(Fast Enrichment Score or Fisher Enrichment Score) uses a different method
of calculating the enrichment score as compared with GSEA.  In FES, the
hypergeometric distribution is used to determine the p-value for the
partitioning of members of the geneset within the ordered valued gene
ranking.  Two different p-values are calculated from the opposite ends of the
ranking to measure enrichment at either the high or low ends of the spectrum.
The use of the hypergeometric distribution rather than the complicated method
used in GSEA means this method is very fast and yields results which are
similar to GSEA itself.  In comparison to GSEA, FES is more sensitive to
concentration of geneset members in the middle of the ranking as happens with
GSEA v1 and is avoided by GSEA v2.  All results from FES or GSEA should be
evaluated by looking at the plots and changes in gene expression to evaluate
the feasibility of detecting the shift in the gene set distribution using
experimental tools.  Typically if the changes are smaller than a 0.5 on the
log2FC scale, they will be challenging to detect by methods such as Q-PCR.
}
