% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.java.exec.R
\name{GSEA.java.exec}
\alias{GSEA.java.exec}
\title{Title}
\usage{
GSEA.java.exec(input.ds, input.cls, gs.db, output.directory = "",
  doc.string = "GSEA.analysis", non.interactive.run = F,
  reshuffling.type = "sample.labels", nperm = 1000,
  weighted.score.type = 1, nom.p.val.threshold = -1,
  fwer.p.val.threshold = -1, fdr.q.val.threshold = 0.25, topgs = 20,
  adjust.FDR.q.val = F, gs.size.threshold.min = 15,
  gs.size.threshold.max = 500, reverse.sign = F, preproc.type = 0,
  random.seed = 123456, perm.type = 0, fraction = 1, replace = F,
  save.intermediate.results = F, OLD.GSEA = F,
  use.fast.enrichment.routine = T, metric = "Signal2Noise", contrast = "",
  silent = T)
}
\arguments{
\item{input.ds}{Input gene expression Affy dataset file in RES or GCT format}

\item{input.cls}{Input class vector (phenotype) file in CLS format}

\item{gs.db}{Gene set database in GMT format}

\item{output.directory}{Directory where to store output and results (default: "")}

\item{doc.string}{Documentation string used as a prefix to name result files (default: "GSEA.analysis")}

\item{non.interactive.run}{Run in interactive (i.e. R GUI) or batch (R command line) mode (default: F)}

\item{reshuffling.type}{Type of permutation reshuffling: "sample.labels" or "gene.labels" (default: "sample.labels"}

\item{nperm}{Number of random permutations (default: 1000)}

\item{weighted.score.type}{Enrichment correlation-based weighting: 0=no weight (KS), 1= weigthed, 2 = over-weigthed (default: 1)}

\item{nom.p.val.threshold}{Significance threshold for nominal p-vals for gene sets (default: -1, no thres)}

\item{fwer.p.val.threshold}{Significance threshold for FWER p-vals for gene sets (default: -1, no thres)}

\item{fdr.q.val.threshold}{Significance threshold for FDR q-vals for gene sets (default: 0.25)}

\item{topgs}{Besides those passing test, number of top scoring gene sets used for detailed reports (default: 10)}

\item{adjust.FDR.q.val}{Adjust the FDR q-vals (default: F)}

\item{gs.size.threshold.min}{Minimum size (in genes) for database gene sets to be considered (default: 25)}

\item{gs.size.threshold.max}{Maximum size (in genes) for database gene sets to be considered (default: 500)}

\item{reverse.sign}{Reverse direction of gene list (pos. enrichment becomes negative, etc.) (default: F)}

\item{preproc.type}{Preproc.normalization: 0=none, 1=col(z-score)., 2=col(rank) and row(z-score)., 3=col(rank). (def: 0)}

\item{random.seed}{Random number generator seed. (default: 123456)}

\item{perm.type}{For experts only. Permutation type: 0 = unbalanced, 1 = balanced (default: 0)}

\item{fraction}{For experts only. Subsampling fraction. Set to 1.0 (no resampling) (default: 1.0)}

\item{replace}{For experts only, Resampling mode (replacement or not replacement) (default: F)}

\item{save.intermediate.results}{For experts only, save intermediate results (e.g. matrix of random perm. scores) (default: F)}

\item{OLD.GSEA}{Use original (old) version of GSEA (default: F)}

\item{use.fast.enrichment.routine}{Use faster routine to compute enrichment for random permutations (default: T)}

\item{silent}{}
}
\description{
Title
}
