% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.gsea.R
\docType{data}
\name{gsea.pkg.env}
\alias{gsea.pkg.env}
\title{Title}
\format{An object of class \code{environment} of length 0.}
\usage{
gsea.pkg.env
}
\arguments{
\item{gsea}{}
}
\description{
Title
}
\keyword{datasets}
