% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{loadSig}
\alias{loadSig}
\title{Loads signatures from GMT formatted files from msigdb}
\usage{
loadSig(filename, path = ".")
}
\arguments{
\item{filename}{The GMT file to read}

\item{path}{The path to the GMT file}
}
\value{
A data frame of the signatures, signature class, definitions, and genes.
}
\description{
Loads signatures from GMT formatted files from msigdb
}
\examples{
sigs <- loadSigs("test.gmt", ".")
}
